pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-news"
version = "0.1.0"
description = "AI News aggregator and newsletter pipeline powered by Claude Agent SDK"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "markdown",
    "requests",
    "python-dotenv",
    "msal",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
ai-news = "ai_news.__main__:main"
